+++ src/Message/OptionsBuilder.php
@@ -8,8 +8,6 @@ use ReflectionClass;
 /**
  * Builder for creation of options used by FCM.
  *
- * Class OptionsBuilder
- *
  * @link http://firebase.google.com/docs/cloud-messaging/http-server-ref#downstream-http-messages-json
  */
 class OptionsBuilder
@@ -51,7 +49,7 @@ class OptionsBuilder
     /**
      * @internal
      *
-     * @var string
+     * @var int
      */
     protected $timeToLive;
 
@@ -70,6 +68,13 @@ class OptionsBuilder
     protected $dryRun = false;
 
     /**
+     * @internal
+     *
+     * @var bool
+     */
+    protected $directBootOk;
+
+    /**
      * This parameter identifies a group of messages
      * A maximum of 4 different collapse keys is allowed at any given time.
      *
@@ -80,6 +85,23 @@ class OptionsBuilder
     public function setCollapseKey($collapseKey)
     {
         $this->collapseKey = $collapseKey;
+
+        return $this;
+    }
+
+    /**
+     * If set to true, messages will be allowed to be delivered to the app while the device is in direct boot mode.
+     * See Support Direct Boot mode (https://developer.android.com/training/articles/direct-boot).
+     *
+     * @see https://developer.android.com/training/articles/direct-boot
+     *
+     * @param true $directBootOk (only true is valid, do not use for false it is pointless)
+     *
+     * @return \LaravelFCM\Message\OptionsBuilder
+     */
+    public function setDirectBootOk($directBootOk)
+    {
+        $this->directBootOk = $directBootOk;
 
         return $this;
     }
@@ -130,7 +152,7 @@ class OptionsBuilder
      * When a notification is sent and this is set to true,
      * the content of the notification can be modified before it is displayed.
      *
-     * @param String $isMutableContent
+     * @param bool $isMutableContent
      * @return OptionsBuilder
      */
     public function setMutableContent($isMutableContent)
@@ -175,6 +197,7 @@ class OptionsBuilder
 
     /**
      * This parameter specifies the package name of the application where the registration tokens must match in order to receive the message.
+     * (Android only)
      *
      * @param string $restrictedPackageName
      *
@@ -191,6 +214,7 @@ class OptionsBuilder
      * This parameter, when set to true, allows developers to test a request without actually sending a message.
      * It should only be used for the development.
      *
+     * @deprecated v1 (https://stackoverflow.com/a/53885050/5155484)
      * @param bool $isDryRun
      *
      * @return \LaravelFCM\Message\OptionsBuilder
@@ -283,6 +307,16 @@ class OptionsBuilder
     }
 
     /**
+     * is direct boot ok
+     *
+     * @return bool
+     */
+    public function isDirectBootOk()
+    {
+        return $this->directBootOk;
+    }
+
+    /**
      * build an instance of Options.
      *
      * @return Options
@@ -293,9 +327,6 @@ class OptionsBuilder
     }
 }
 
-/**
- * Class OptionsPriorities.
- */
 final class OptionsPriorities
 {
     /**
@@ -323,7 +354,7 @@ final class OptionsPriorities
     /**
      * check if this priority is supported by fcm.
      *
-     * @param $priority
+     * @param string $priority
      *
      * @return bool
      *