+++ src/Request/Request.php
@@ -7,9 +7,6 @@ use LaravelFCM\Message\Options;
 use LaravelFCM\Message\PayloadData;
 use LaravelFCM\Message\PayloadNotification;
 
-/**
- * Class Request.
- */
 class Request extends BaseRequest
 {
     /**
@@ -50,7 +47,7 @@ class Request extends BaseRequest
     /**
      * Request constructor.
      *
-     * @param                     $to
+     * @param string|array        $to
      * @param Options             $options
      * @param PayloadNotification $notification
      * @param PayloadData         $data