+++ src/Sender/FCMSender.php
@@ -12,9 +12,6 @@ use GuzzleHttp\Exception\ClientException;
 use LaravelFCM\Response\DownstreamResponse;
 use LaravelFCM\Message\PayloadNotification;
 
-/**
- * Class FCMSender.
- */
 class FCMSender extends HTTPSender
 {
     const MAX_TOKEN_PER_REQUEST = 1000;
@@ -43,7 +40,7 @@ class FCMSender extends HTTPSender
 
                 $responseGuzzle = $this->post($request);
 
-                $responsePartial = new DownstreamResponse($responseGuzzle, $tokens);
+                $responsePartial = new DownstreamResponse($responseGuzzle, $tokens, $this->logger);
                 if (!$response) {
                     $response = $responsePartial;
                 } else {
@@ -54,7 +51,7 @@ class FCMSender extends HTTPSender
             $request = new Request($to, $options, $notification, $data);
             $responseGuzzle = $this->post($request);
 
-            $response = new DownstreamResponse($responseGuzzle, $to);
+            $response = new DownstreamResponse($responseGuzzle, $to, $this->logger);
         }
 
         return $response;
@@ -63,7 +60,7 @@ class FCMSender extends HTTPSender
     /**
      * Send a message to a group of devices identified with them notification key.
      *
-     * @param                          $notificationKey
+     * @param string|string[]          $notificationKey
      * @param Options|null             $options
      * @param PayloadNotification|null $notification
      * @param PayloadData|null         $data
@@ -76,7 +73,7 @@ class FCMSender extends HTTPSender
 
         $responseGuzzle = $this->post($request);
 
-        return new GroupResponse($responseGuzzle, $notificationKey);
+        return new GroupResponse($responseGuzzle, $notificationKey, $this->logger);
     }
 
     /**
@@ -95,7 +92,7 @@ class FCMSender extends HTTPSender
 
         $responseGuzzle = $this->post($request);
 
-        return new TopicResponse($responseGuzzle, $topics);
+        return new TopicResponse($responseGuzzle, $topics, $this->logger);
     }
 
     /**