+++ tests/TopicsTest.php
@@ -5,13 +5,11 @@ use GuzzleHttp\Psr7\Response;
 use LaravelFCM\Message\Topics;
 use LaravelFCM\Sender\FCMSender;
 use LaravelFCM\Message\Exceptions\NoTopicProvidedException;
+use LaravelFCM\Sender\FCMTopic;
 
 class TopicsTest extends FCMTestCase
 {
-    /**
-     * @test
-     */
-    public function it_throw_an_exception_if_no_topic_is_provided()
+    public function testItThrowAnExceptionIfNoTopicIsProvided()
     {
         $topics = new Topics();
 
@@ -19,10 +17,7 @@ class TopicsTest extends FCMTestCase
         $topics->build();
     }
 
-    /**
-     * @test
-     */
-    public function it_has_only_one_topic()
+    public function testItHasOnlyOneTopic()
     {
         $target = '/topics/myTopic';
 
@@ -33,10 +28,7 @@ class TopicsTest extends FCMTestCase
         $this->assertEquals($target, $topics->build());
     }
 
-    /**
-     * @test
-     */
-    public function it_has_two_topics_and()
+    public function testItHasTwoTopicsAnd()
     {
         $target = [
             'condition' => "'firstTopic' in topics && 'secondTopic' in topics",
@@ -49,10 +41,7 @@ class TopicsTest extends FCMTestCase
         $this->assertEquals($target, $topics->build());
     }
 
-    /**
-     * @test
-     */
-    public function it_has_two_topics_or()
+    public function testItHasTwoTopicsOr()
     {
         $target = [
             'condition' => "'firstTopic' in topics || 'secondTopic' in topics",
@@ -65,10 +54,7 @@ class TopicsTest extends FCMTestCase
         $this->assertEquals($target, $topics->build());
     }
 
-    /**
-     * @test
-     */
-    public function it_has_two_topics_or_and_one_and()
+    public function testItHasTwoTopicsOrAndOneAnd()
     {
         $target = [
             'condition' => "'firstTopic' in topics || 'secondTopic' in topics && 'thirdTopic' in topics",
@@ -81,10 +67,7 @@ class TopicsTest extends FCMTestCase
         $this->assertEquals($target, $topics->build());
     }
 
-    /**
-     * @test
-     */
-    public function it_has_a_complex_topic_condition()
+    public function testItHasAComplexTopicCondition()
     {
         $target = [
             'condition' => "'TopicA' in topics && ('TopicB' in topics || 'TopicC' in topics) || ('TopicD' in topics && 'TopicE' in topics)",
@@ -103,17 +86,17 @@ class TopicsTest extends FCMTestCase
         $this->assertEquals($target, $topics->build());
     }
 
-    /**
-     * @test
-     */
-    public function it_send_a_notification_to_a_topic()
+    public function testItSendsANotificationToATopic()
     {
         $response = new Response(200, [], '{"message_id":6177433633397011933}');
 
         $client = Mockery::mock(Client::class);
         $client->shouldReceive('request')->once()->andReturn($response);
 
-        $fcm = new FCMSender($client, 'http://test.test');
+        $logger = new \Monolog\Logger('test');
+        $logger->pushHandler(new \Monolog\Handler\NullHandler());
+
+        $fcm = new FCMSender($client, 'http://test.test', $logger);
 
         $topics = new Topics();
         $topics->topic('test');
@@ -125,17 +108,17 @@ class TopicsTest extends FCMTestCase
         $this->assertNull($response->error());
     }
 
-    /**
-     * @test
-     */
-    public function it_send_a_notification_to_a_topic_and_return_error()
+    public function testItSendsANotificationToATopicAndReturnError()
     {
         $response = new Response(200, [], '{"error":"TopicsMessageRateExceeded"}');
 
         $client = Mockery::mock(Client::class);
         $client->shouldReceive('request')->once()->andReturn($response);
 
-        $fcm = new FCMSender($client, 'http://test.test');
+        $logger = new \Monolog\Logger('test');
+        $logger->pushHandler(new \Monolog\Handler\NullHandler());
+
+        $fcm = new FCMSender($client, 'http://test.test', $logger);
 
         $topics = new Topics();
         $topics->topic('test');
@@ -145,5 +128,85 @@ class TopicsTest extends FCMTestCase
         $this->assertFalse($response->isSuccess());
         $this->assertTrue($response->shouldRetry());
         $this->assertEquals('TopicsMessageRateExceeded', $response->error());
+    }
+
+    public function testCreateTopic()
+    {
+        $response = new Response(200, [], json_encode([]));
+
+        $client = Mockery::mock(Client::class);
+        $client->shouldReceive('request')->once()->andReturn($response);
+
+        $logger = new \Monolog\Logger('test');
+        $logger->pushHandler(new \Monolog\Handler\NullHandler());
+
+        $fcm = new FCMTopic($client, 'http://test.test', $logger);
+
+        $response = $fcm->createTopic('id_1', 'abcd');
+        $this->assertTrue($response);
+    }
+
+    public function testSubscribeTopic()
+    {
+        $response = new Response(200, [], json_encode([]));
+
+        $client = Mockery::mock(Client::class);
+        $client->shouldReceive('request')->once()->andReturn($response);
+
+        $logger = new \Monolog\Logger('test');
+        $logger->pushHandler(new \Monolog\Handler\NullHandler());
+
+        $fcm = new FCMTopic($client, 'http://test.test', $logger);
+
+        $response = $fcm->subscribeTopic('id_1', 'token1');
+        $this->assertTrue($response);
+    }
+
+    public function testSubscribeTopicMultipleTokens()
+    {
+        $response = new Response(200, [], json_encode([]));
+
+        $client = Mockery::mock(Client::class);
+        $client->shouldReceive('request')->once()->andReturn($response);
+
+        $logger = new \Monolog\Logger('test');
+        $logger->pushHandler(new \Monolog\Handler\NullHandler());
+
+        $fcm = new FCMTopic($client, 'http://test.test', $logger);
+
+        $response = $fcm->subscribeTopic('id_1', ['token1', 'token2']);
+        $this->assertTrue($response);
+    }
+
+    public function testUnSubscribeTopic()
+    {
+        $response = new Response(200, [], json_encode([]));
+
+        $client = Mockery::mock(Client::class);
+        $client->shouldReceive('request')->once()->andReturn($response);
+
+        $logger = new \Monolog\Logger('test');
+        $logger->pushHandler(new \Monolog\Handler\NullHandler());
+
+        $fcm = new FCMTopic($client, 'http://test.test', $logger);
+
+        $response = $fcm->unSubscribeTopic('id_1', 'token1');
+        $this->assertTrue($response);
+    }
+
+    public function testUnSubscribeTopicMultipleTokens()
+    {
+        $response = new Response(200, [], json_encode([]));
+
+        $client = Mockery::mock(Client::class);
+        $client->shouldReceive('request')->once()->andReturn($response);
+
+        $logger = new \Monolog\Logger('test');
+        $logger->pushHandler(new \Monolog\Handler\NullHandler());
+
+        $fcm = new FCMTopic($client, 'http://test.test', $logger);
+
+        $response = $fcm->unSubscribeTopic('id_1', ['token1', 'token2']);
+        $this->assertTrue($response);
     }
 }